requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0

# LangChain + OpenAI
langchain==0.3.14
langchain-openai==0.3.0
openai==1.58.1

# Utilities
pydantic==2.10.4
python-dotenv==1.0.1
